// frontend/scratch/old-src/frontend/src/shared/lib/types.ts
export type PageResponse<T> = {
  content: T[];
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
};
// ...
export type ThreadResponse = {
  id: number;
  title: string;
  slug: string;
  categoryName: string;
  categorySlug: string;
  authorUsername: string;
  replyCount: number;
  viewCount: number;
  pinned: boolean;
  locked: boolean;
  hidden: boolean;
  liked: boolean;
  likes: number;
  createdAt: string;
  lastPostAt: string;
  tags?: string[];
};

export type PostResponse = {
  id: number;
  content: string;
  authorUsername: string;
  liked: boolean;
  likes: number;
  hidden: boolean;
  createdAt: string;
  bestAnswer: boolean;
};

export type ThreadDetailResponse = {
  thread: ThreadResponse;
  posts: PageResponse<PostResponse>;
};